LayoutTests/inspector/editor/text-editor-char-to-coordinates-expected.txt

Test editor cursorPositionToCoordinates and coordinatesToCursorPosition API


Running: testCoordinatesToCursorPositionOuter
Request char at (-1000, -1000)
Result:null

Running: testTextToCoordinatesCornerCases
Request (-1, 0):null
Request (100, 0):null
Request (0, -1):null
Request (0, 100):null
Request (-100, -1100):null

Running: testInverseRelationTextToCoordinates
Testing line 'function foo(a, b) {'
Testing line '    var f = /*.[a]/.test(a);'
Testing line '    return f;'
